[minsik] add map click and hide list

+++ app/src/main/java/com/example/dbproject/data/ReviewData.kt
@@ -4,10 +4,10 @@ data class ReviewData(
     var resName : String? = null,
     var resId : Int? = null,
     var userName: String? = null,
-    var resRating: Float? = 0.0f,
+    var resRating: Double? = 0.0,
     var favorite: Int? = 0,
     var comment: String? = null,
     var imageUrl: String? = null,
     var favoriteList: ArrayList<String> = arrayListOf<String>(),
-    var menuRating: ArrayList<Float> = arrayListOf<Float>()
+    var menuRating: ArrayList<Double> = arrayListOf<Double>()
 )
